requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
gradio>=4.0.0
pymupdf>=1.23.0
pdfplumber>=0.10.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
transformers>=4.35.0
torch>=2.0.0
FlagEmbedding>=1.2.0
groq>=0.4.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pymongo>=4.6.0
resend>=2.0.0
